Bug fix: Use slice notation to copy the requests.
When a list has been successfully started on DPM, the `activeReqs` field is assigned the value of `stagedReqs`. This is incorrect because now both fields hold a reference to the same underlying object. If the application modifies `stagedReqs`, `activeReqs` will also apear to have been modified.

This commit changes the assignment to a copy so the application can freely change `stagedReqs` without affecting the current data returns.